Feat/typescript

+++ contracts/interfaces/ICurve.sol
@@ -1,13 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.0;
 
-
 /// @title Curve Interface
 /// @author Carl Farterson (@carlfarterson)
 /// @dev Required for all Curves
 interface ICurve {
-
-    event Updated(uint indexed hubId);
+    event Updated(uint256 indexed hubId);
 
     /// @notice Given a hub, baseX, baseY and connector weight, add the configuration to the
     /// BancorZero ValueSet registry
@@ -15,46 +13,41 @@ interface ICurve {
     ///      that may contain different ValueSet arguments
     /// @param _hubId                   unique hub identifier
     /// @param _encodedValueSet     encoded ValueSet arguments
-    function register(
-        uint _hubId,
-        bytes calldata _encodedValueSet
-    ) external;
+    function register(uint256 _hubId, bytes calldata _encodedValueSet) external;
 
     /// @notice TODO
     /// @param _hubId                   unique hub identifier
     /// @param _encodedValueSet     encoded target ValueSet arguments
-    function registerTarget(
-        uint _hubId,
-        bytes calldata _encodedValueSet
-    ) external;
+    function registerTarget(uint256 _hubId, bytes calldata _encodedValueSet)
+        external;
 
     function calculateMintReturn(
-        uint _tokensDeposited,
-        uint _hubId,
-        uint _supply,
-        uint _balancePooled
-    ) external view returns (uint meTokensReturned);
+        uint256 _tokensDeposited,
+        uint256 _hubId,
+        uint256 _supply,
+        uint256 _balancePooled
+    ) external view returns (uint256 meTokensReturned);
 
     function calculateBurnReturn(
-        uint _meTokensBurned,
-        uint _hubId,
-        uint _supply,
-        uint _balancePooled
-    ) external view returns (uint tokensReturned);
+        uint256 _meTokensBurned,
+        uint256 _hubId,
+        uint256 _supply,
+        uint256 _balancePooled
+    ) external view returns (uint256 tokensReturned);
 
     function calculateTargetMintReturn(
-        uint _tokensDeposited,
-        uint _hubId,
-        uint _supply,
-        uint _balancePooled
-    ) external view returns (uint meTokensReturned);
+        uint256 _tokensDeposited,
+        uint256 _hubId,
+        uint256 _supply,
+        uint256 _balancePooled
+    ) external view returns (uint256 meTokensReturned);
 
     function calculateTargetBurnReturn(
-        uint _meTokensBurned,
-        uint _hubId,
-        uint _supply,
-        uint _balancePooled
-    ) external view returns (uint tokensReturned);
+        uint256 _meTokensBurned,
+        uint256 _hubId,
+        uint256 _supply,
+        uint256 _balancePooled
+    ) external view returns (uint256 tokensReturned);
 
-    function finishUpdate(uint id) external;
+    function finishUpdate(uint256 id) external;
 }
