+++ contracts/libs/Details.sol
@@ -1,43 +1,38 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity ^0.8.0;
+
 library Details {
     struct MeTokenDetails {
         address owner;
         uint256 hubId;
         uint256 balancePooled;
         uint256 balanceLocked;
-        
         bool updating; // TODO: validate
-        uint startTime;
-        uint endTime;
-
-        uint targetHub;
+        uint256 startTime;
+        uint256 endTime;
+        uint256 targetHub;
     }
 
     struct HubDetails {
         bool active;
-
         address vault;
         address curve;
-        uint refundRatio;
-        
+        uint256 refundRatio;
         bool updating;
-        uint startTime;
-        uint endTime;
-
+        uint256 startTime;
+        uint256 endTime;
         address migrationVault;
         address targetVault;
         bool curveDetails;
         address targetCurve;
-        uint targetRefundRatio;
+        uint256 targetRefundRatio;
     }
 
     struct BancorDetails {
-        uint baseY;
+        uint256 baseY;
         uint32 reserveWeight;
-
         // bool updating;
-        uint targetBaseY;
+        uint256 targetBaseY;
         uint32 targetReserveWeight;
     }
 }
